requirements.txt
adbutils==2.12.0
alembic==1.18.1
androguard==4.1.3
annotated-types==0.7.0
anyio==4.12.0
apkInspector==1.3.6
asn1crypto==1.5.1
asttokens==3.0.1
banal==1.0.6
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
dataset==1.6.2
decorator==5.2.1
deprecation==2.1.0
distro==1.9.0
dotenv==0.9.9
executing==2.2.1
filelock==3.20.3
fonttools==4.61.1
frida==17.6.1
fsspec==2026.1.0
greenlet==3.3.0
h11==0.16.0
hmdriver2==1.4.4
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
ImageHash==4.3.2
ipython==9.9.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.9
langchain==1.2.6
langchain-core==1.2.7
langgraph==1.0.7
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.7
langgraph-sdk==0.3.3
langsmith==0.6.4
loguru==0.7.3
lxml==5.4.0
Mako==1.3.10
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mpmath==1.3.0
mutf8==1.0.6
networkx==3.6.1
numpy==2.4.1
openai==2.13.0
opencv-python==4.13.0.90
orjson==3.11.5
ormsgpack==1.12.2
packaging==25.0
parso==0.8.5
pillow==12.1.0
prompt_toolkit==3.0.52
pure_eval==0.2.3
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydot==4.0.1
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyWavelets==1.9.0
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
requests-toolbelt==1.0.0
retry2==0.9.5
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.0
sentence-transformers==5.2.0
setuptools==80.10.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==1.4.54
stack-data==0.6.3
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.10.0
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.6
typing-inspection==0.4.2
typing_extensions==4.15.0
uiautomator2==3.5.0
urllib3==2.6.3
uuid_utils==0.14.0
wcwidth==0.3.1
win32_setctime==1.2.0
xxhash==3.6.0
zstandard==0.25.0
